new bot handlers architecture
