+++ jest.config.js
@@ -13,4 +13,5 @@ export default {
       },
     ],
   },
+  setupFilesAfterEnv: ['<rootDir>/src/jest.setup.ts'],
 };
