+++ src/bot/bot-router.service.ts
@@ -1,7 +1,11 @@
 import { Telegraf } from 'telegraf';
 import { message } from 'telegraf/filters';
-import type { BotService } from './bot.service.ts';
 import logger from '../shared/logger/logger.ts';
+import type { AddMedicationHandler } from './handlers/add-medication.handler.ts';
+import type { AddMedicationTextHandler } from './handlers/add-medication-text-handler.ts';
+import { OnStartHandler } from './handlers/on-start.handler.ts';
+import type { OnHelpHandler } from './handlers/on-help.handler.ts';
+import type { OnCancelHandler } from './handlers/on-cancel.handler.ts';
 
 export class BotRouterService {
   private readonly commands = [
@@ -12,7 +16,13 @@ export class BotRouterService {
     { command: 'contact_developer', description: 'Contact the developer' },
   ];
 
-  constructor(private readonly botService: BotService) {}
+  constructor(
+    private readonly onStartHandler: OnStartHandler,
+    private readonly onHelpHandler: OnHelpHandler,
+    private readonly addMedicationHandler: AddMedicationHandler,
+    private readonly textHandler: AddMedicationTextHandler,
+    private readonly onCancelHandler: OnCancelHandler
+  ) {}
 
   /**
    * Sets up command handlers for the bot
@@ -21,15 +31,15 @@ export class BotRouterService {
   async setupCommands(bot: Telegraf): Promise<void> {
     logger.info('Bot setup commands', { commands: this.commands });
     await bot.telegram.setMyCommands(this.commands);
-    await this.setupNotificationHandlers(bot);
+    await this.setupNotificationHandlers(/*bot*/);
 
-    bot.start(ctx => this.botService.onStart(ctx));
+    bot.start(ctx => this.onStartHandler.handle(ctx));
 
-    bot.help(ctx => this.botService.onHelp(ctx));
+    bot.help(ctx => this.onHelpHandler.handle(ctx));
 
-    bot.command('add_medication', async ctx => this.botService.onAddMedication(ctx));
+    bot.command('add_medication', async ctx => this.addMedicationHandler.handle(ctx));
 
-    bot.command('cancel_add_medication', async ctx => this.botService.onCancelAddMedication(ctx));
+    bot.command('cancel_add_medication', async ctx => this.onCancelHandler.handle(ctx));
 
     // Write to developer command
     bot.command('contact_developer', ctx => {
@@ -38,12 +48,12 @@ export class BotRouterService {
       );
     });
 
-    bot.on(message('text'), async ctx => this.botService.onText(ctx));
+    bot.on(message('text'), async ctx => this.textHandler.handle(ctx));
   }
 
-  private async setupNotificationHandlers(bot: Telegraf): Promise<void> {
-    bot.action(/skip_(\d+)/, async ctx => this.botService.onSkipNotification(ctx));
-
-    bot.action(/accept_(\d+)/, async ctx => this.botService.onAcceptNotification(ctx));
+  private async setupNotificationHandlers(/*bot: Telegraf*/): Promise<void> {
+    // todo call separate handelrs
+    // bot.action(/skip_(\d+)/, async ctx => this.botService.onSkipNotification(ctx));
+    // bot.action(/accept_(\d+)/, async ctx => this.botService.onAcceptNotification(ctx));
   }
 }
