+++ src/bot/types/bot-command-handler.interface.ts
@@ -1,0 +1,5 @@
+import type { BotContext, TextMessage } from './context.type.ts';
+
+export interface BotCommandHandler<C extends BotContext> {
+  handle(ctx: C): Promise<TextMessage>;
+}
