+++ src/bot/types/context.type.ts
@@ -1,11 +1,15 @@
 import type { Context, NarrowedContext } from 'telegraf';
-import type { Message, Update } from 'telegraf/typings/core/types/typegram';
+import type { Update } from 'telegraf/typings/core/types/typegram';
+import type * as tg from 'telegraf/types';
 
-export type TextMessageContext = NarrowedContext<
-  Context<Update>,
-  Update.MessageUpdate<Message.TextMessage>
->;
+export type TextMessage = tg.Message.TextMessage;
 
-export type ActionContext = NarrowedContext<Context, Update.CallbackQueryUpdate> & {
+export interface BotContext extends Context<Update> {
+  match?: RegExpMatchArray;
+}
+
+export type TextMessageContext = NarrowedContext<BotContext, Update.MessageUpdate<TextMessage>>;
+
+export type ActionContext = NarrowedContext<BotContext, Update.CallbackQueryUpdate> & {
   match: RegExpMatchArray;
 };
